Add IL Tree interpreter
Adding the basic post-order traversal interpretation mechanism of the
IL tree. Right now the traversal is only done for each tree top in one
basic block.

This interpreter is implemented as one of the optimizations
in order to directly interpret the IL before any other optimization
happens.

A nodeValueMap is used to save the results of nodes as they are processed for
commoning. A symbol table is also kept for storing variables for load
and store operations.

Signed-off-by: Alex Li <tianyi7@ualberta.ca>